# Add additional hooks for subclasses