Fixes #22

I see nti.site doing something like this in its subclass:
```python
class _RunJobInSite(TransactionLoop):

    def setUp(self):
        db = component.getUtility(IDatabase)
        self.conn = db.open()
        self.tearDown = self.conn.close
```

That means the object can't be used from multiple threads at the same time, but it's not used in that way anyway.


Also some minor refactoring with an eye towards speed and simplicity.

We get between 5 and 10% faster, even accounting for all the overhead of raising exceptions:

| Benchmark      | master |  this branch               |
|----------------|---------------|------------------------------|
| trivial commit | 14.9 us       | 13.2 us: 1.12x faster (-11%) |
| max retries    | 16.0 us       | 15.1 us: 1.06x faster (-5%)  |