@jamadden 2019-09-06
I need to release this project in order to make the necessary changes to nti.site. But I'd *really* like to see https://github.com/zopefoundation/transaction/pull/84 merged and released before releasing this so that we can fix #17 and remove one of [the XXX comments](https://github.com/NextThought/nti.transactions/blob/43fbfddb9a988634c746e5a74c8ce441691bd8ad/src/nti/transactions/transactions.py#L584) that explicit transactions are designed to help with in the first place.