@jzuech3 2019-09-06
I'd advocate we release this project as-is to get as much alpha burn in time as possible for these changes, coming back later for the transaction changes.