@jamadden 2019-09-06
That's a fair point. The aborthook thing is pretty minor, and I think I've thought of a way to handle the XXX thing in a safe-enough way even without the transaction update; it's mostly a hypothetical issue anyway.